@@ -3,19 +3,22 @@ package br.com.wesjon
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
-import kotlinx.android.synthetic.main.activity_main.*
+import br.com.wesjon.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+    lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         setupUpExampleList()
     }
 
     private fun setupUpExampleList() {
-        sample_list.adapter = AnimationSampleAdapter(
+        binding.sampleList.adapter = AnimationSampleAdapter(
             object : AnimationSampleListener {
                 override fun onClick(animationSample: AnimationSamples) {
                     startActivity(AnimationCanvasActivity.getIntent(this@MainActivity, animationSample))
@@ -24,7 +27,7 @@ class MainActivity : AppCompatActivity() {
         ).apply {
             submitList(AnimationSamples.values().toList())
         }
-        sample_list.addItemDecoration(DividerItemDecoration(this, DividerItemDecoration.VERTICAL))
+        binding.sampleList.addItemDecoration(DividerItemDecoration(this, DividerItemDecoration.VERTICAL))
     }
 }
 
